+++ src/components/Details.tsx
@@ -111,6 +111,8 @@ export default function Details({ file }: DetailsProps) {
             return detail.showAudio;
           case Protocol.StreamKind.Text:
             return detail.showSubtitle;
+          case Protocol.StreamKind.Menu:
+            return detail.showMenu;
           default:
             return true;
         }